refactor(commands): reimplement run for compatibility
Reimplements the run command to be compatible with Shards. The previous implementation
required changing the scripts to a mapping of the different targets which Shards does
not support. This version takes a new approach by keeping the original format but
allowing platform targets to be included following the script name and the '@' symbol.